fix: bugs fix

+++ src/migration/v2_0_1/secret.py
@@ -15,17 +15,17 @@ _LOGGER = logging.getLogger(DEFAULT_LOGGER)
 def secret_secret_migration(
     mongo_client: MongoCustomClient, domain_id_param, project_map
 ):
-    schema_id = ""
-    resource_group = "PROJECT"
-    project_id = ""
-    workspace_id = ""
-
     for secret_info in mongo_client.find(
         "SECRET", "secret", {"domain_id": domain_id_param}, {}
     ):
         if secret_info.get("workspace_id"):
             continue
-
+        
+        schema_id = ''
+        resource_group = "PROJECT"
+        project_id = ''
+        workspace_id = ''
+        
         schema = secret_info.get("schema")
         if schema:
             schema_id = _get_schema_to_schema_id(schema)
@@ -34,6 +34,7 @@ def secret_secret_migration(
             workspace_id = project_map[secret_info["domain_id"]].get(
                 secret_info.get("project_id")
             )
+            project_id = secret_info.get("project_id")
         elif secret_info.get("service_account_id"):
             service_account_info = mongo_client.find_one(
                 "IDENTITY",
@@ -54,12 +55,10 @@ def secret_secret_migration(
 
             workspace_id = service_account_info.get("workspace_id")
             project_id = service_account_info.get("project_id")
+        else:
+            resource_group = "DOMAIN"
+            workspace_id = "*"
             
-        if not workspace_id:
-            _LOGGER.error(
-                f"Secret service_account({secret_info['service_account_id']}) does not exists in IDENTITY.sa"
-            )
-        
         set_param = {
             "$set": {
                 "secret_schema_id": schema_id,
@@ -136,43 +135,44 @@ def secret_trusted_secret_migration(mongo_client: MongoCustomClient, domain_id_p
                 , {"service_account_id": trusted_secret_info.get("service_account_id")}, {}
             )
 
-            new_trusted_account_id = generate_id("ta")
-            trusted_account_create = {
-                "trusted_account_id": new_trusted_account_id,
-                "name": service_account_info.get("name"),
-                "data": service_account_info.get("data"),
-                "provider": service_account_info.get("provider"),
-                "tags": service_account_info.get("tags"),
-                "secret_schema_id": schema_id,
-                "trusted_secret_id": trusted_secret_info.get("trusted_secret_id"),
-                "resource_group": "DOMAIN",
-                "workspace_id": "*",
-                "domain_id": service_account_info["domain_id"],
-                "created_at": datetime.utcnow(),
-            }
-
-            mongo_client.insert_one(
-                "IDENTITY", "trusted_account", trusted_account_create
-            )
-
-            mongo_client.update_one(
-                "SECRET", "trusted_secret"
-                , {"_id": trusted_secret_info["_id"]}
-                , {
-                    "$set": {"trusted_account_id": new_trusted_account_id},
-                    "$unset": {"service_account_id": 1}
+            if service_account_info:
+                new_trusted_account_id = generate_id("ta")
+                trusted_account_create = {
+                    "trusted_account_id": new_trusted_account_id,
+                    "name": service_account_info.get("name"),
+                    "data": service_account_info.get("data"),
+                    "provider": service_account_info.get("provider"),
+                    "tags": service_account_info.get("tags"),
+                    "secret_schema_id": schema_id,
+                    "trusted_secret_id": trusted_secret_info.get("trusted_secret_id"),
+                    "resource_group": "DOMAIN",
+                    "workspace_id": "*",
+                    "domain_id": service_account_info["domain_id"],
+                    "created_at": datetime.utcnow(),
                 }
-            )
 
-            mongo_client.delete_many(
-                "IDENTITY", "service_account"
-                , {"service_account_id": service_account_info["service_account_id"]}
-            )
+                mongo_client.insert_one(
+                    "IDENTITY", "trusted_account", trusted_account_create
+                )
+
+                mongo_client.update_one(
+                    "SECRET", "trusted_secret"
+                    , {"_id": trusted_secret_info["_id"]}
+                    , {
+                        "$set": {"trusted_account_id": new_trusted_account_id},
+                        "$unset": {"service_account_id": 1}
+                    }
+                )
+
+                mongo_client.delete_many(
+                    "IDENTITY", "service_account"
+                    , {"service_account_id": service_account_info["service_account_id"]}
+                )
 
 
 def _get_schema_to_schema_id(schema):
     
-    schema_id = None
+    schema_id = schema
     if schema == "azure_subscription_id":
         schema_id = "azure-secret-subscription-id"
     elif schema == "azure_client_secret":
